Уведомления в Telegram при аутентификации и выходе

+++ app/Http/Controllers/Api/V1/Admin/AuthController.php
@@ -9,54 +9,45 @@ use App\Http\Controllers\ApiController;
 use App\Http\Requests\User\LoginRequest;
 use App\Http\Resources\V1\User\UserResource;
 use App\Models\User;
-use App\Notifications\LoginAdminNotification;
+use Illuminate\Auth\Events\Attempting;
+use Illuminate\Auth\Events\Login;
 use Illuminate\Auth\Events\Logout;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Symfony\Component\HttpFoundation\Response;
 
 final class AuthController extends ApiController
 {
     public function login(LoginRequest $request)
-    : JsonResponse
-    {
+    : JsonResponse {
         $credentials = $request->validated();
 
-        if (Auth::attempt([
-            'email'     => $credentials['email'],
-            'password'  => $credentials['password'],
-            'is_active' => true,
-            'is_admin'  => true,
-        ],
-        )) {
-            $role_ids = DB::table('roles')
-                ->where(
-                    'slug',
-                    '!=',
-                    RoleEnum::MANAGER->getName(),
-                )
-                ->pluck('id');
+        event(new Attempting('admin', $credentials, false));
 
-            $admin = User::query()
-                ->where('email', $credentials['email'])
-                ->where('is_admin', true)
-                ->whereIn('role_id', $role_ids)
-                ->first();
+        $role_ids = DB::table('roles')
+            ->where(
+                'slug',
+                '!=',
+                RoleEnum::MANAGER->getName(),
+            )
+            ->pluck('id');
 
+        $admin = User::query()
+            ->where('email', $credentials['email'])
+            ->where('is_admin', true)
+            ->whereIn('role_id', $role_ids)
+            ->first();
+
+        if ($admin) {
             $role = $admin->role->slug;
-
             $admin->tokens()->delete();
-
             $token = $admin
                 ->createToken("Token for $role: $admin->display_name", ['*'], now()->addHours(24))
                 ->plainTextToken;
-
             $admin->update([
                 'logged_in_at' => now(),
             ]);
-
-            $admin->notify(new LoginAdminNotification());
+            event(new Login('admin', $admin, false));
 
             return $this->successResponse(
                 ['token' => $token],
@@ -85,7 +76,9 @@ final class AuthController extends ApiController
         $admin = auth()->user();
         $admin->tokens()->delete();
 
-        event(new Logout('web', $admin));
+        DB::delete('delete from sessions where user_id = ?', [$admin->id]);
+
+        event(new Logout('admin', $admin));
 
         return $this->successResponse(
             '',
